requirements.txt
scikit-learn~=1.2.2
matplotlib~=3.7.1
numpy~=1.23.5
tqdm==4.66.1
torch~=2.0.0
torchvision~=0.15.1
seaborn~=0.12.2
opencv-contrib-python~=4.7.0.72
mlflow~=2.5.0
Pillow~=9.4.0
scipy~=1.11.1
setuptools~=65.5.1
segmentation-models-pytorch~=0.3.3
jupyter~=1.0.0
labelme~=5.3.1
